backend/requirements.txt
fastapi
uvicorn
groq
python-dotenv
reportlab
pydantic
